+++ src/compressor/onpair16.rs
@@ -5,7 +5,7 @@ use rustc_hash::FxHashMap;
 use rand::seq::SliceRandom;
 use rand::thread_rng;
 
-const THRESHOLD: usize = 10; 
+const THRESHOLD: u16 = 10; 
 const MAX_LENGTH: usize = 16;
 
 pub struct OnPair16Compressor {
